+++ yggdrasil_decision_forests/port/python/ydf/model/generic_model.py
@@ -364,7 +364,9 @@ Use `model.describe()` for more details
   def evaluate(
       self,
       data: dataset.InputDataset,
+      *,
       bootstrapping: Union[bool, int] = False,
+      weighted: bool = False,
   ) -> metric.Evaluation:
     """Evaluates the quality of a model on a dataset.
 
@@ -400,6 +402,9 @@ Use `model.describe()` for more details
         to an integer, it specifies the number of bootstrapping samples to use.
         In this case, if the number is less than 100, an error is raised as
         bootstrapping will not yield useful results.
+      weighted: If true, the evaluation is weighted according to the training
+        weights. If false, the evaluation is non-weighted. b/351279797: Change
+        default to weights=True.
 
     Returns:
       Model evaluation.
@@ -426,7 +431,9 @@ Use `model.describe()` for more details
           task=self.task()._to_proto_type(),  # pylint: disable=protected-access
       )
 
-      evaluation_proto = self._model.Evaluate(ds._dataset, options_proto)  # pylint: disable=protected-access
+      evaluation_proto = self._model.Evaluate(
+          ds._dataset, options_proto, weighted=weighted
+      )  # pylint: disable=protected-access
     return metric.Evaluation(evaluation_proto)
 
   def analyze_prediction(
@@ -616,6 +623,8 @@ Use `model.describe()` for more details
       pre_processing: Optional[Callable] = None,  # pylint: disable=g-bare-generic
       post_processing: Optional[Callable] = None,  # pylint: disable=g-bare-generic
       temp_dir: Optional[str] = None,
+      tensor_specs: Optional[Dict[str, Any]] = None,
+      feature_specs: Optional[Dict[str, Any]] = None,
   ) -> None:
     """Exports the model as a TensorFlow Saved model.
 
@@ -673,25 +682,97 @@ Use `model.describe()` for more details
     )
     ```
 
-    The SavedModel format allows for custom preprocessing and postprocessing
-    computation in addition to the model inference. Such computation can be
-    specified with the `pre_processing` and `post_processing` arguments:
+    Some TensorFlow Serving or Servomatic pipelines rely on feed examples as
+    serialized TensorFlow Example proto (instead of raw tensor values) and/or
+    wrap the model raw output (e.g. probability predictions) into a special
+    structure (called the Serving API). You can create models compatible with
+    those two convensions with `feed_example_proto=True` and `servo_api=True`
+    respectively:
 
     ```python
-    def pre_processing(features):
-      features = features.copy()
-      features["f1"] = features["f1"] * 2
-      return features
-
     model.to_tensorflow_saved_model(
         path="/tmp/my_model",
         mode="tf",
-        pre_processing=pre_processing,
+        feed_example_proto=True,
+        servo_api=True
     )
     ```
 
-    For more complex combinations, such as composing multiple models, use the
-    method `to_tensorflow_function` instead of `to_tensorflow_saved_model`.
+    If your model requires some data preprocessing or post-processing, you can
+    express them as a @tf.function or a tf module and pass them to the
+    `pre_processing` and `post_processing` arguments respectively.
+
+    Warning: When exporting a SavedModel, YDF infers the model signature using
+    the dtype of the features observed during training. If the signature of the
+    pre_processing function is different than the signature of the model (e.g.,
+    the processing creates a new feature), you need to specify the tensor specs
+    (`tensor_specs`; if `feed_example_proto=False`) or feature spec
+    (`feature_specs`; if `feed_example_proto=True`) argument:
+
+    ```python
+    # Define a pre-processing function
+    @tf.function
+    def pre_processing(raw_features):
+      features = {**raw_features}
+      # Create a new feature.
+      features["sin_f1"] = tf.sin(features["f1"])
+      # Remove a feature
+      del features["f1"]
+      return features
+
+    # Create Numpy dataset
+    raw_dataset = {
+        "f1": np.random.random(size=100),
+        "f2": np.random.random(size=100),
+        "l": np.random.randint(2, size=100),
+    }
+
+    # Apply the preprocessing on the training dataset.
+    processed_dataset = (
+        tf.data.Dataset.from_tensor_slices(raw_dataset)
+        .batch(128)  # The batch size has no impact on the model.
+        .map(preprocessing)
+        .prefetch(tf.data.AUTOTUNE)
+    )
+
+    # Train a model on the pre-processed dataset.
+    ydf_model = specialized_learners.RandomForestLearner(
+        label="l",
+        task=generic_learner.Task.CLASSIFICATION,
+    ).train(processed_dataset)
+
+    # Export the model to a raw SavedModel model with the pre-processing
+    model.to_tensorflow_saved_model(
+        path="/tmp/my_model",
+        mode="tf",
+        feed_example_proto=False,
+        pre_processing=pre_processing,
+        tensor_specs{
+            "f1": tf.TensorSpec(shape=[None], name="f1", dtype=tf.float64),
+            "f2": tf.TensorSpec(shape=[None], name="f2", dtype=tf.float64),
+        }
+    )
+
+    # Export the model to a SavedModel consuming serialized tf examples with the
+    # pre-processing
+    model.to_tensorflow_saved_model(
+        path="/tmp/my_model",
+        mode="tf",
+        feed_example_proto=True,
+        pre_processing=pre_processing,
+        feature_specs={
+            "f1": tf.io.FixedLenFeature(
+                shape=[], dtype=tf.float32, default_value=math.nan
+            ),
+            "f2": tf.io.FixedLenFeature(
+                shape=[], dtype=tf.float32, default_value=math.nan
+            ),
+        }
+    )
+    ```
+
+    For more flexibility, use the method `to_tensorflow_function` instead of
+    `to_tensorflow_saved_model`.
 
     Args:
       path: Path to store the Tensorflow Decision Forests model.
@@ -712,8 +793,10 @@ Use `model.describe()` for more details
       feature_dtypes: Mapping from feature name to TensorFlow dtype. Use this
         mapping to feature dtype. For instance, numerical features are encoded
         with tf.float32 by default. If you plan on feeding tf.float64 or
-        tf.int32, use `feature_dtype` to specify it. Only compatible with
-        mode="tf".
+        tf.int32, use `feature_dtype` to specify it. `feature_dtypes` is ignored
+        if `tensor_specs` is set. If set, disables the automatic signature
+        extraction on `pre_processing` (if `pre_processing` is also set). Only
+        compatible with mode="tf".
       servo_api: If true, adds a SavedModel signature to make the model
         compatible with the `Classify` or `Regress` servo APIs. Only compatible
         with mode="tf". If false, outputs the raw model predictions.
@@ -723,12 +806,32 @@ Use `model.describe()` for more details
         provided as a binary serialized TensorFlow Example proto. This is the
         format expected by VertexAI and most TensorFlow Serving pipelines.
       pre_processing: Optional TensorFlow function or module to apply on the
-        input features before applying the model. Only compatible with
-        mode="tf".
+        input features before applying the model. If the `pre_processing`
+        function has been traced (i.e., the function has been called once with
+        actual data and contains a concrete instance in its cache), this
+        signature is extracted and used as signature of the SavedModel. Only
+        compatible with mode="tf".
       post_processing: Optional TensorFlow function or module to apply on the
         model predictions. Only compatible with mode="tf".
       temp_dir: Temporary directory used during the conversion. If None
         (default), uses `tempfile.mkdtemp` default temporary directory.
+      tensor_specs: Optional dictionary of `tf.TensorSpec` that define the input
+        features of the model to export. If not provided, the TensorSpecs are
+        automatically generated based on the model features seen during
+        training. This means that "tensor_specs" is only necessary when using a
+        "pre_processing" argument that expects different features than what the
+        model was trained with. This argument is ignored when exporting model
+        with `feed_example_proto=True` as in this case, the TensorSpecs are
+        defined by the `tf.io.parse_example` parsing feature specs. Only
+        compatible with mode="tf".
+      feature_specs: Optional dictionary of `tf.io.parse_example` parsing
+        feature specs e.g. `tf.io.FixedLenFeature` or `tf.io.RaggedFeature`. If
+        not provided, the praising feature specs are automatically generated
+        based on the model features seen during training. This means that
+        "feature_specs" is only necessary when using a "pre_processing" argument
+        that expects different features than what the model was trained with.
+        This argument is ignored when exporting model with
+        `feed_example_proto=False`. Only compatible with mode="tf".
     """
 
     if mode == "keras":
@@ -751,6 +854,8 @@ Use `model.describe()` for more details
         pre_processing=pre_processing,
         post_processing=post_processing,
         temp_dir=temp_dir,
+        tensor_specs=tensor_specs,
+        feature_specs=feature_specs,
     )
 
   def to_tensorflow_function(  # pytype: disable=name-error
@@ -1095,7 +1200,11 @@ Use `model.describe()` for more details
     self._model.ForceEngine(engine_name)
 
 
-def from_sklearn(sklearn_model: Any) -> GenericModel:
+def from_sklearn(
+    sklearn_model: Any,
+    label_name: str = "label",
+    feature_name: str = "features",
+) -> GenericModel:
   """Converts a tree-based scikit-learn model to a YDF model.
 
   Usage example:
@@ -1129,17 +1238,28 @@ def from_sklearn(sklearn_model: Any) -> GenericModel:
   *   sklearn.ensemble.ExtraTreesClassifier
   *   sklearn.ensemble.ExtraTreesRegressor
   *   sklearn.ensemble.GradientBoostingRegressor
+  *   sklearn.ensemble.IsolationForest
+
+  Unlike YDF, Scikit-learn does not name features and labels. Use the fields
+  `label_name` and `feature_name` to specify the name of the columns in the YDF
+  model.
 
   Additionally, only single-label classification and scalar regression are
   supported (e.g. multivariate regression models will not convert).
 
   Args:
     sklearn_model: the scikit-learn tree based model to be converted.
+    label_name: Name of the multi-dimensional feature in the output YDF model.
+    feature_name: Name of the label in the output YDF model.
 
   Returns:
     a YDF Model that emulates the provided scikit-learn model.
   """
-  return _get_export_sklearn().from_sklearn(sklearn_model)
+  return _get_export_sklearn().from_sklearn(
+      sklearn_model=sklearn_model,
+      label_name=label_name,
+      feature_name=feature_name,
+  )
 
 
 def _get_export_jax():